+++ R/cols_add.R
@@ -384,10 +384,6 @@ cols_add <- function(
     resolved_column_before <- NULL
   } else if (length(resolved_column_before) != 1) {
 
-    if (length(resolved_column_before) < 1) {
-      cli::cli_abort("The expression used for `.before` did not resolve a column.")
-    }
-
     if (length(resolved_column_before) > 1) {
       cli::cli_abort("The expression used for `.before` resolved multiple columns.")
     }
@@ -404,14 +400,15 @@ cols_add <- function(
     resolved_column_after <- NULL
   } else if (length(resolved_column_after) != 1) {
 
-    if (length(resolved_column_after) < 1) {
-      cli::cli_abort("The expression used for `.after` did not resolve a column.")
-    }
-
     if (length(resolved_column_after) > 1) {
       cli::cli_abort("The expression used for `.after` resolved multiple columns.")
     }
   }
+  
+  if (length(resolved_column_after) == 1 && resolved_column_after == colnames(data_tbl)[NCOL(data_tbl)]) {
+    # if requesting the last column to add after, use NULL instead.
+    resolved_column_after <- NULL
+  } 
 
   # Stop function if expressions are given to both `.before` and `.after`
   if (!is.null(resolved_column_before) && !is.null(resolved_column_after)) {
@@ -455,9 +452,9 @@ cols_add <- function(
     } else {
       updated_data_tbl <-
         dplyr::bind_cols(
-          dplyr::select(data_tbl, 1:(before_colnum - 1)),
+          dplyr::select(data_tbl, 1:(dplyr::all_of(before_colnum) - 1)),
           data_tbl_new_cols,
-          dplyr::select(data_tbl, before_colnum:ncol(data_tbl))
+          dplyr::select(data_tbl, dplyr::all_of(before_colnum):ncol(data_tbl))
         )
     }
 
@@ -473,7 +470,7 @@ cols_add <- function(
 
   } else if (is.null(resolved_column_before) && !is.null(resolved_column_after)) {
 
-    if (resolved_column_after == nrow(data_tbl)) {
+    if (resolved_column_after == ncol(data_tbl)) {
 
       updated_data_tbl <-
         vctrs::vec_cbind(
@@ -493,14 +490,14 @@ cols_add <- function(
 
       if (after_colnum >= ncol(data_tbl)) {
         updated_data_tbl <-
-          dplyr::bind_cols(
+          vctrs::vec_cbind(
             data_tbl,
             data_tbl_new_cols
           )
       } else {
         updated_data_tbl <-
           dplyr::bind_cols(
-            dplyr::select(data_tbl, 1:after_colnum),
+            dplyr::select(data_tbl, 1:dplyr::all_of(after_colnum)),
             data_tbl_new_cols,
             dplyr::select(data_tbl, (after_colnum + 1):ncol(data_tbl))
           )
@@ -508,7 +505,7 @@ cols_add <- function(
 
 
       after_colnum <- which(boxh_df[["var"]] == resolved_column_after)
-
+      
       updated_boxh_df <-
         dplyr::bind_rows(
           boxh_df[1:after_colnum, ],